src/backend/decision-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aio-pika==9.4.0
redis==5.0.1
httpx==0.26.0
python-dotenv==1.0.0
